% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compile_model}
\alias{compile_model}
\title{Compile a Stan model and return a path to the compiled model output.}
\usage{
compile_model(model_file)
}
\arguments{
\item{model_file}{Path to a Stan model file.
This is a text file with the model spceification.}
}
\value{
Path to the compiled Stan model, which is just an RDS file.
  To run the model, you can read this file into a new R session with
  `readRDS()` and feed it to the `object` argument of `sampling()`.
}
\description{
We return the paths to the Stan model specification
  and the compiled model file so `targets` can treat them as
  dynamic files and compile the model if either file changes.
}
\examples{
library(cmdstanr)
compile_model("stan/model.stan")
}
\references{
https://github.com/wlandau/targets-stan
}
