% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_sim_model}
\alias{fit_sim_model}
\title{Fit the Stan model}
\usage{
fit_sim_model(
  data,
  model_file,
  iter_warmup = 1,
  iter_sampling = 1,
  adapt_delta = 0.9,
  max_treedepth = 15,
  chains = 4,
  parallel_chains = 1,
  refresh = 0,
  seed = 123
)
}
\arguments{
\item{data}{list, a single simulated dataset.}

\item{model_file}{Path to the Stan model source file.}
}
\value{
dataframe of cmdstan customized summary with diagnostics
}
\description{
Fit the Stan model
}
\references{
https://github.com/wlandau/targets-stan
}
