% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_para}
\alias{get_para}
\title{Get posterior estimates mcmc summary}
\usage{
get_para(data, row, col, digits = 2, nsmall = 2)
}
\arguments{
\item{data}{data frame, summary of mcmc}

\item{row}{variable name (e.g., "theta")}

\item{col}{summary name (e.g., "mean", "q50")}

\item{digits}{integer indicating the number of decimal places}

\item{nsmall}{the minimum number of digits to the right of the decimal point}
}
\description{
Get posterior estimates mcmc summary
}
