% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_ggsave}
\alias{my_ggsave}
\title{ggsave for targets}
\usage{
my_ggsave(
  filename,
  plot,
  units = c("in", "cm", "mm", "px"),
  height = NA,
  width = NA,
  dpi = 300,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
ggsave for targets
}
