% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_write_csv}
\alias{my_write_csv}
\title{write_csv for targets}
\usage{
my_write_csv(x, path, append = FALSE, col_names = !append)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{path}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the \code{file} argument
instead.}

\item{append}{If \code{FALSE}, will overwrite existing file. If \code{TRUE},
will append to existing file. In both cases, if the file does not exist a new
file is created.}

\item{col_names}{If \code{FALSE}, column names will not be included at the top of the file. If \code{TRUE},
column names will be included. If not specified, \code{col_names} will take the opposite value given to \code{append}.}
}
\description{
write_csv for targets
}
