% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quiet}
\alias{quiet}
\title{Suppress output and messages for code.}
\usage{
quiet(code)
}
\arguments{
\item{code}{Code to run quietly.}
}
\value{
The result of running the code.
}
\description{
Used in the pipeline.
}
\examples{
library(cmdstanr)
library(tidyverse)
compile_model("stan/model.stan")
quiet(fit_model("stan/model.stan", simulate_data_discrete()))
out
}
\references{
https://github.com/wlandau/targets-stan
}
