% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_schools_data}
\alias{simulate_schools_data}
\title{Simulate data from the model.}
\usage{
simulate_schools_data(tau = 5)
}
\arguments{
\item{tau}{tau parameter for the model.}
}
\description{
Simulate data from the model.
}
